requirements.txt
asgiref==3.3.4
autopep8==1.5.6
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==2.2
django-allauth==0.44.0
django-braces==1.14.0
django-crispy-forms==1.7.0
django-heroku==0.3.1
django-messages==0.6.0
django-model-utils==4.1.1
django-phonenumber-field==5.2.0
django-postman==4.2
django-private-chat==0.3.0
django-storages==1.11.1
dropbox==11.10.0
gunicorn==20.1.0
idna==2.10
oauthlib==3.1.1
phonenumbers==8.12.22
pillow==8.2.0
ply==3.11
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2017.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stone==3.2.1
toml==0.10.2
urllib3==1.26.5
websockets==8.1
whitenoise==5.2.0
